refactor: split confluence-client into html-to-markdown and macro-converter modules
Extract pure HTML-to-markdown conversion into lib/html-to-markdown.js and
Confluence storage-format macro conversion into lib/macro-converter.js.
Reduces lib/confluence-client.js from 2281 to 1738 lines and clarifies the
module boundary between API/auth concerns and content conversion.

- lib/html-to-markdown.js: pure function, no state, zero dependencies
- lib/macro-converter.js: MacroConverter class with { isCloud, webUrlPrefix,
  buildUrl } dependency injection
- ConfluenceClient keeps thin delegating wrappers for backward compatibility
- createLocalConverter() now returns a MacroConverter instance

No behavior change. All 236 existing tests pass.
